refactor(memory-core): split dreaming-phases.ts into 3 focused files
Split the 1828-line dreaming-phases.ts into three focused modules:
- dreaming-shared-types.ts: shared types, constants, normalize utils
- dreaming-session-corpus.ts: session transcript scanning logic
- dreaming-phases.ts: daily ingestion + phase runners + sweep orchestrator

Reduces largest single file from 1828 to 983 lines.
